Move `HeaderBreadcrumbs` component
